+++ Cargo.toml
@@ -61,6 +61,7 @@ uv-settings = { path = "crates/uv-settings" }
 uv-shell = { path = "crates/uv-shell" }
 uv-state = { path = "crates/uv-state" }
 uv-static = { path = "crates/uv-static" }
+uv-trampoline-builder = { path = "crates/uv-trampoline-builder" }
 uv-tool = { path = "crates/uv-tool" }
 uv-types = { path = "crates/uv-types" }
 uv-version = { path = "crates/uv-version" }
