+++ crates/uv-trampoline/rust-toolchain.toml
@@ -1,2 +1,2 @@
 [toolchain]
-channel = "nightly-2024-06-08"
+channel = "nightly-2024-10-27"
